requirements.txt
django
djangorestframework
psycopg
psycopg2
gunicorn
whitenoise
django-heroku
django-cors-headers